chore(*): Refactor and apply review comments.
